Move sqlite db into persistence data layer
