+++ minitwit_nodejs/persistence/models/put_models_here
@@ -1,0 +1,1 @@
+Models used by ORM and by possibly several different database implementations.
